Fix config load order
